pyproject.toml
[project]
name = "jira-agent"
version = "0.1.0"
license = "Apache-2.0"
description = "jira agent"
authors = [
    { name = "AGNTCY Internet of Agents Collective", email = "pypi@agntcy.io" },
]
maintainers = [
    { name = "AGNTCY Internet of Agents Collective", email = "pypi@agntcy.io" },
]
requires-python = ">=3.11,<4.0"
dependencies = [
    "agentevals (==0.0.5)",
    "aiohappyeyeballs (==2.6.1)",
    "aiohttp (==3.11.14)",
    "aiosignal (==1.3.2)",
    "annotated-types (==0.7.0)",
    "anyio (==4.9.0)",
    "attrs (==25.3.0)",
    "build (==1.2.2.post1)",
    "cachecontrol (==0.14.2)",
    "certifi (==2025.1.31)",
    "cffi (==1.17.1)",
    "charset-normalizer (==3.4.1)",
    "cleo (==2.1.0)",
    "click (==8.1.8)",
    "coloredlogs (==15.0.1)",
    "crashtest (==0.4.1)",
    "cryptography (==43.0.3)",
    "defusedxml (==0.7.1)",
    "distlib (==0.3.9)",
    "distro (==1.9.0)",
    "dulwich (==0.22.8)",
    "fastapi (==0.115.11)",
    "fastapi-cli (==0.0.7)",
    "fastjsonschema (==2.21.1)",
    "filelock (==3.18.0)",
    "findpython (==0.6.3)",
    "frozenlist (==1.5.0)",
    "h11 (==0.14.0)",
    "httpcore (==1.0.7)",
    "httptools (==0.6.4)",
    "httpx (==0.28.1)",
    "humanfriendly (==10.0)",
    "idna (==3.10)",
    "iniconfig (==2.1.0)",
    "installer (==0.7.0)",
    "jaraco-classes (==3.4.0)",
    "jaraco-context (==6.0.1)",
    "jaraco-functools (==4.1.0)",
    "jira (==3.8.0)",
    "jiter (==0.9.0)",
    "jsonpatch (==1.33)",
    "jsonpointer (==3.0.0)",
    "jsonschema-rs (==0.20.0)",
    "keyring (==25.6.0)",
    "langchain (>=0.3.23,<0.4.0)",
    "langchain-core (>=0.3.54,<0.4.0)",
    "langchain-openai (==0.3.6)",
    "langchain-text-splitters (>=0.3.8,<0.4.0)",
    "langgraph (>=0.3.31,<0.4.0)",
    "langgraph-api (>=0.1.9,<0.2.0)",
    "langgraph-checkpoint (>=2.0.24,<3.0.0)",
    "langgraph-cli (>=0.2.5,<0.3.0)",
    "langgraph-prebuilt (>=0.1.8,<0.2.0)",
    "langgraph-sdk (>=0.1.61,<0.2.0)",
    "langgraph-supervisor (==0.0.9)",
    "langsmith (>=0.3.32,<0.4.0)",
    "markdown-it-py (==3.0.0)",
    "mdurl (==0.1.2)",
    "more-itertools (==10.6.0)",
    "msgpack (==1.1.0)",
    "multidict (==6.2.0)",
    "numpy (==2.2.4)",
    "oauthlib (==3.2.2)",
    "openai (==1.66.5)",
    "openevals (>=0.0.19,<0.0.20)",
    "orjson (==3.10.15)",
    "packaging (==24.2)",
    "pbs-installer (==2025.4.9)",
    "pillow (==11.1.0)",
    "pkginfo (==1.12.1.2)",
    "platformdirs (==4.3.7)",
    "pluggy (==1.5.0)",
    "poetry (==2.1.2)",
    "poetry-core (==2.1.2)",
    "propcache (==0.3.0)",
    "pycparser (==2.22)",
    "pydantic (==2.10.6)",
    "pydantic-settings (==2.8.1)",
    "pydantic-core (==2.27.2)",
    "pygments (==2.19.1)",
    "pyjwt (==2.10.1)",
    "pyproject-hooks (==1.2.0)",
    "pytest (==7.4.4)",
    "python-dotenv (==1.0.1)",
    "pyyaml (==6.0.2)",
    "rapidfuzz (==3.13.0)",
    "regex (==2024.11.6)",
    "requests (==2.32.3)",
    "requests-oauthlib (==2.0.0)",
    "requests-toolbelt (==1.0.0)",
    "rich (==13.9.4)",
    "rich-toolkit (==0.13.2)",
    "ruff (==0.0.289)",
    "setuptools (==78.1.0)",
    "shellingham (==1.5.4)",
    "sniffio (==1.3.1)",
    "sqlalchemy (==2.0.39)",
    "sse-starlette (==2.1.3)",
    "starlette (==0.46.1)",
    "structlog (==25.2.0)",
    "tabulate (==0.9.0)",
    "tenacity (==9.0.0)",
    "tiktoken (==0.9.0)",
    "tomlkit (==0.13.2)",
    "tqdm (==4.67.1)",
    "trove-classifiers (==2025.4.11.15)",
    "typer (==0.15.2)",
    "typing-extensions (==4.12.2)",
    "urllib3 (==2.3.0)",
    "uvicorn (==0.34.0)",
    "uvloop (==0.21.0)",
    "virtualenv (==20.30.0)",
    "watchfiles (==1.0.4)",
    "websockets (==15.0.1)",
    "xattr (==1.1.4)",
    "yarl (==1.18.3)",
    "zstandard (==0.23.0)",
    "cisco-outshift-agent-utils (==0.0.2)",
    "langchain-anthropic (>=0.3.12,<0.4.0)",
    "agntcy-acp (==1.1.2)",
    "httpx-sse (>=0.4.0,<0.5.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
